[TEST] Move tests to main for conftest detection
For some reason (figuring it out) submodule tests can't detect
conftest.py from above.
