+++ abagen/datasets/__init__.py
@@ -8,6 +8,6 @@ __all__ = [
     'fetch_gene_group', 'WELL_KNOWN_IDS', '_get_dataset_dir'
 ]
 
-from .datasets import (fetch_microarray, fetch_raw_mri, fetch_desikan_killiany,
+from .fetchers import (fetch_microarray, fetch_raw_mri, fetch_desikan_killiany,
                        fetch_gene_group, WELL_KNOWN_IDS)
 from .utils import _get_dataset_dir
